line app/requirements.txt
line-bot-sdk==3.21.0
Flask==3.1.0
APScheduler==3.11.0
